Move menu components into new dir
